refactor: 학생 다운로드 util 함수 구조 개선 및 엑셀 형식 수정
