+++ apps/graduate/src/shared/utils/studentDownload/index.ts
@@ -1,3 +1,2 @@
 export { default as downloadStudentDetailExcel } from './downloadStudentDetailExcel';
-export { default as downloadStudentDetailFromRows } from './downloadStudentDetailFromRows';
 export { default as handleDownload } from './handleDownload';
